+++ ReleaseNotes.md
@@ -3,6 +3,7 @@ F# niggles part 2
 * Explicitly exempt the `get_` and `set_` prefixes of getter and setter methods from the `AvoidNonAlphanumericIdentifierRule` since that was not already a thing.
 * Make `AvoidUnneededUnboxingRule` not applicable to [CompilerGenerated] functions (e.g. Union case `CompareTo`)
 * Exempt debugger-related generated types related to union types from `AvoidUnsealedUninheritedInternalTypeRule`
+* Exempt getter and setter methods from `ConsiderConvertingMethodToPropertyRule` (well, duh!)
 
 For previous releases, go here -- https://github.com/SteveGilham/Gendarme/blob/master/ReleaseNotes%20-%20Previously.md
 Since this is a release from a fork, [issues should be reported at my related repo](https://github.com/SteveGilham/altcode.fake/issues) that contains a Fake driver for the Gendarme tool, but noted as being against the forked Gendarme tool itself.